[Game] Fix instanced transfer loading
