--- AAEmu.Game/Models/Tasks/Transfers/TransferTickStartTask.cs
@@ -1,11 +1,0 @@
-﻿using AAEmu.Game.Core.Managers;
-
-namespace AAEmu.Game.Models.Tasks.Transfers;
-
-public class TransferTickStartTask : Task
-{
-    public override void Execute()
-    {
-        TransferManager.Instance.TransferTick();
-    }
-}
